refactor: split main.rs into modules (types, utils, capture, handlers, websocket, tests)
- Move types/constants/shared aliases to src/types.rs
- Move parsing/validation and local IP helper to src/utils.rs
- Extract tshark capture manager/tasks to src/capture.rs
- Extract HTTP handlers to src/handlers.rs
- Extract WebSocket handler/task to src/websocket.rs
- Move tests to src/tests.rs and wire #[cfg(test)] in main.rs

No behavior changes; API and WS protocol unchanged.

Testing:
- cargo build --verbose: OK
- cargo test --verbose: 3 passed; 0 failed
- cargo fmt -- --check: rustfmt not installed (skipped)
